Make retoreToFirstTimer also work from inside chooseNodeMongoApiDBServer  function
